Update patch set 1
Patch Set 1:

Hi Vijay- I agree that there is no need to regenerate these POJO classes as published VES specification is not going to change. 

My thinking was that not to touch the classes that are generated and not written. If you don't agree with this, I am ok to accept this change.

Patch-set: 1
Reviewer: Gerrit User 233 <233@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0